feat(hyperliquid): fold outcomes scraper into the hyperliquid service
When substreams-hypercore consolidated the outcome carve-out back
into one package (v0.4.0), the scraper still ran as two separate
services / two pods / two cycles. Folding them mirrors the
substreams story: one HL Info client per env, one polling cadence,
one set of CH credentials, one /live + /metrics endpoint.

Restructured services/hyperliquid/ into two pure-function sub-cycles
(`runSpotCycle`, `runOutcomesCycle`) orchestrated by `run()` in
`index.ts` via `Promise.allSettled`. Failure in either surfaces to
the supervisor without short-circuiting the other.

Removed:
- services/hyperliquid-outcomes/ (deleted entirely)
- `hyperliquid-outcomes` SERVICE registration in cli.ts
- `setup hyperliquid-outcomes` CLI command (folded into
  `setup hyperliquid` which now deploys all three tables)

Tests restructured: each sub-cycle gets its own test suite over the
pure cycle function; new `index.test.ts` covers the orchestrator's
parallel-execution + error-surfacing semantics. End-to-end validated
against dev1 v0.4.2 — single cycle inserts 307 spot pair names + 138
outcomes + 22 questions in ~1.1s.
