+++ services/hyperliquid/outcomes.test.ts
@@ -9,13 +9,8 @@ const mockQuery = mock(() =>
 );
 const mockIncrementSuccess = mock(() => {});
 const mockIncrementError = mock(() => {});
-const mockInitService = mock(() => {});
 const mockMarkServiceAlive = mock(() => {});
 
-// Mock only the `lib/clickhouse` exports this service imports
-// (`insertClient` + `query`). `mock.module` is process-wide, so providing
-// fields beyond what's needed would risk shadowing real exports for other
-// suites.
 mock.module('../../lib/clickhouse', () => ({
     insertClient: { insert: mockInsert },
     query: mockQuery,
@@ -26,8 +21,11 @@ mock.module('../../lib/prometheus', () => ({
     incrementError: mockIncrementError,
 }));
 
+// `mock.module` applies process-wide for the test session. Mirror the full set
+// of exports other test files mock so a later `mock.module` call in another
+// suite can't leave dangling undefined imports for whichever test runs last.
 mock.module('../../lib/service-init', () => ({
-    initService: mockInitService,
+    initService: mock(() => {}),
     markServiceAlive: mockMarkServiceAlive,
 }));
 
@@ -78,14 +76,6 @@ interface InsertCallArg {
     format: string;
 }
 
-/**
- * Build a `mockQuery` implementation that routes to different result sets
- * depending on which CH table the SQL references. Tests need this because
- * the service fires two reads per cycle — `outcome_fills` for the known
- * universe and `state_outcome_meta` for already-settled ids — and feeding
- * both the same rows would mark every known id as already-settled and
- * suppress the settled-lookup probes the test is verifying.
- */
 function mockQueryRouter(routes: {
     knownIds?: string[];
     alreadySettled?: string[];
@@ -110,16 +100,14 @@ function mockQueryRouter(routes: {
     };
 }
 
-describe('hyperliquid-outcomes run()', () => {
+describe('hyperliquid runOutcomesCycle()', () => {
     const originalFetch = globalThis.fetch;
-    const originalUrl = process.env.HYPERLIQUID_INFO_URL;
 
     beforeEach(() => {
         mockInsert.mockClear();
         mockQuery.mockClear();
         mockIncrementSuccess.mockClear();
         mockIncrementError.mockClear();
-        mockInitService.mockClear();
         mockMarkServiceAlive.mockClear();
         mockQuery.mockImplementation(() =>
             Promise.resolve({
@@ -135,24 +123,9 @@ describe('hyperliquid-outcomes run()', () => {
 
     afterEach(() => {
         globalThis.fetch = originalFetch;
-        if (originalUrl === undefined) {
-            delete process.env.HYPERLIQUID_INFO_URL;
-        } else {
-            process.env.HYPERLIQUID_INFO_URL = originalUrl;
-        }
-    });
-
-    test('throws when HYPERLIQUID_INFO_URL is unset', async () => {
-        delete process.env.HYPERLIQUID_INFO_URL;
-        const { run } = await import('./index');
-        await expect(run()).rejects.toThrow(/HYPERLIQUID_INFO_URL/);
     });
 
     test('inserts live + settled outcomes and questions in one cycle', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
-        // outcome_fills knows 104, 172, and a settled 0 — 0 is missing from
-        // the live snapshot and not in already-settled, so the cycle must
-        // probe settledOutcome for it.
         mockQuery.mockImplementation(
             mockQueryRouter({
                 knownIds: ['104', '172', '0'],
@@ -175,8 +148,8 @@ describe('hyperliquid-outcomes run()', () => {
             return Promise.resolve(new Response('null', { status: 200 }));
         }) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await run();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await runOutcomesCycle('http://example/info');
 
         expect(mockInsert).toHaveBeenCalledTimes(2);
         const calls = mockInsert.mock.calls.map((c) => c[0] as InsertCallArg);
@@ -189,7 +162,6 @@ describe('hyperliquid-outcomes run()', () => {
         expect(questionCall).toBeDefined();
         expect(outcomeCall!.format).toBe('JSONEachRow');
 
-        // 2 live + 1 settled
         expect(outcomeCall!.values).toHaveLength(3);
         const live104 = outcomeCall!.values.find((r) => r.outcome_id === 104);
         const settled0 = outcomeCall!.values.find((r) => r.outcome_id === 0);
@@ -198,9 +170,7 @@ describe('hyperliquid-outcomes run()', () => {
         expect(settled0?.status).toBe('settled');
         expect(settled0?.settle_fraction).toBe(0);
         expect(settled0?.settle_details).toBe('price:78212.4');
-        // 172 belongs to question 32 via namedOutcomes — reverse map must apply.
         expect(live172?.question_id).toBe(32);
-        // 104 is standalone.
         expect(live104?.question_id).toBeNull();
 
         expect(questionCall!.values).toHaveLength(1);
@@ -208,13 +178,10 @@ describe('hyperliquid-outcomes run()', () => {
 
         expect(mockIncrementSuccess).toHaveBeenCalledTimes(1);
         expect(mockIncrementError).not.toHaveBeenCalled();
-        // Direct-insert services bump the wall-clock heartbeat so the
-        // liveness probe reflects progress (batch queue is unused here).
         expect(mockMarkServiceAlive).toHaveBeenCalledTimes(1);
     });
 
     test('proceeds when known-ids query fails (cold cluster)', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
         mockQuery.mockImplementation(() =>
             Promise.reject(new Error('table not found')),
         );
@@ -224,17 +191,15 @@ describe('hyperliquid-outcomes run()', () => {
             ),
         ) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await run();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await runOutcomesCycle('http://example/info');
 
-        // Still inserted live outcomes + questions.
         expect(mockInsert).toHaveBeenCalledTimes(2);
         expect(mockIncrementSuccess).toHaveBeenCalledTimes(1);
         expect(mockIncrementError).not.toHaveBeenCalled();
     });
 
     test('continues past per-id settledOutcome failures without aborting the cycle', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
         mockQuery.mockImplementation(
             mockQueryRouter({
                 knownIds: ['999'],
@@ -248,28 +213,21 @@ describe('hyperliquid-outcomes run()', () => {
                     new Response(JSON.stringify(liveBody), { status: 200 }),
                 );
             }
-            // settledOutcome lookups fail — cycle should still complete with
-            // the live snapshot.
             return Promise.resolve(new Response('boom', { status: 502 }));
         }) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await run();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await runOutcomesCycle('http://example/info');
 
         expect(mockInsert).toHaveBeenCalledTimes(2);
         const outcomeCall = mockInsert.mock.calls
             .map((c) => c[0] as InsertCallArg)
             .find((c) => c.table === 'state_outcome_meta');
-        // Only the 2 live outcomes — settled lookup failed and was skipped.
         expect(outcomeCall!.values).toHaveLength(2);
         expect(mockIncrementSuccess).toHaveBeenCalledTimes(1);
     });
 
     test('skips settledOutcome probes for ids already captured as settled', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
-        // outcome_fills has 104, 172, and a historically-settled 0; but our
-        // own `state_outcome_meta` already contains 0 with status='settled',
-        // so the cycle must NOT re-probe HL for it.
         mockQuery.mockImplementation(
             mockQueryRouter({
                 knownIds: ['104', '172', '0'],
@@ -290,22 +248,19 @@ describe('hyperliquid-outcomes run()', () => {
             return Promise.resolve(new Response('null', { status: 200 }));
         }) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await run();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await runOutcomesCycle('http://example/info');
 
         expect(settledProbed).toEqual([]);
         const outcomeCall = mockInsert.mock.calls
             .map((c) => c[0] as InsertCallArg)
             .find((c) => c.table === 'state_outcome_meta');
-        // Only the 2 live outcomes inserted this cycle — settled 0 stayed in
-        // the table from the prior cycle's insert (RMT keeps the older row).
         expect(outcomeCall!.values).toHaveLength(2);
         expect(mockIncrementSuccess).toHaveBeenCalledTimes(1);
         expect(mockMarkServiceAlive).toHaveBeenCalledTimes(1);
     });
 
     test('returns early without success metric when outcomeMeta is empty', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
         globalThis.fetch = mock(() =>
             Promise.resolve(
                 new Response(JSON.stringify({ outcomes: [], questions: [] }), {
@@ -314,11 +269,9 @@ describe('hyperliquid-outcomes run()', () => {
             ),
         ) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await run();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await runOutcomesCycle('http://example/info');
 
-        // Empty universe is a soft anomaly — don't insert, don't claim
-        // success, but also don't throw. Next cycle retries.
         expect(mockInsert).not.toHaveBeenCalled();
         expect(mockIncrementSuccess).not.toHaveBeenCalled();
         expect(mockIncrementError).not.toHaveBeenCalled();
@@ -326,17 +279,14 @@ describe('hyperliquid-outcomes run()', () => {
     });
 
     test('records an error metric and rethrows when outcomeMeta fetch fails', async () => {
-        process.env.HYPERLIQUID_INFO_URL = 'http://example/info';
         globalThis.fetch = mock(() =>
             Promise.resolve(new Response('nope', { status: 502 })),
         ) as unknown as typeof fetch;
 
-        const { run } = await import('./index');
-        await expect(run()).rejects.toThrow();
+        const { runOutcomesCycle } = await import('./outcomes');
+        await expect(runOutcomesCycle('http://example/info')).rejects.toThrow();
         expect(mockIncrementError).toHaveBeenCalledTimes(1);
         expect(mockInsert).not.toHaveBeenCalled();
-        // Liveness heartbeat must NOT advance on a failed cycle, so silent
-        // upstream failures still surface via the `/live` probe.
         expect(mockMarkServiceAlive).not.toHaveBeenCalled();
     });
 });
